Update old Redox firmware to latest status quo
Add dedicated colon on Raise for Redox like Kyria

Add volume up/down to the redox keyboard

On the Kyrias this is implemented with the encoder but I don't have that
on the Redox.

Add caps word enabler in Redox rules

I don't believe this should be necessary but at least now it works.

Clean up Redox rules.mk